gameover bounce sound edit

+++ Assets/Scripts/SB/SoundManager.cs
@@ -115,7 +115,7 @@ public class SoundManager : Singleton<SoundManager> {
         aimAudioSource.Play();
     }
 
-    public void ChangeBounceVol(int volume) {
+    public void ChangeBounceVol(float volume) {
         volDict[Sound.Player] = volume;
         volDict[Sound.Enemy1] = volume;
         volDict[Sound.Enemy2] = volume;
